feat: load GPT prompt from file
